pyproject.toml
[project]
name = "mrekiba-bao"
version = "1.0.0" # x-release-please-version
description = "Ansible collection for managing OpenBao via the HTTP API"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [{ name = "Mohamed Rekiba" }]
dependencies = [
    "hvac>=2.4.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.15.7",
    "ansible-lint>=26.3.0",
    "pre-commit>=4.0.0",
]

[project.urls]
Repository = "https://github.com/mohamed-rekiba/ansible-bao"
Issues = "https://github.com/mohamed-rekiba/ansible-bao/issues"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.lint.per-file-ignores]
# Ansible modules must declare DOCUMENTATION/EXAMPLES/RETURN before imports
"plugins/modules/*.py" = ["E402"]
